init work
