+++ src/scripts/UtilitySystem/index.js
@@ -1,0 +1,3 @@
+import * as UtilitySystem from './helpers';
+
+export { UtilitySystem };
